+++ src/index.d.ts
@@ -1,8 +1,17 @@
 import { Writable, Readable, Updater } from 'svelte/store';
 
+export type ParsableStoreNames = 'path' | 'query' | 'fragment';
+
+export type Hook<T> = (
+    value: T | null,
+    currValue: T,
+    storeName: ParsableStoreNames
+) => boolean | undefined;
+
 export interface Parsable<T> extends Readable<T> {
     set(this: void, value: T | string): void;
     update(this: void, updater: Updater<T | string>): void;
+    hook(fn: Hook<T>): () => void;
 }
 
 export interface SubmitEvent extends Event {